requirements.txt
# Core API & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
streamlit==1.30.0

# Image Processing & CV
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
Pillow==10.2.0
pdf2image==1.17.0
PyPDF2==3.0.1
pikepdf==8.11.2
albumentations==1.3.1
scikit-image==0.22.0

# OCR
paddleocr==2.7.0.3
paddlepaddle==2.6.2
pytesseract==0.3.10

# ML & Detection
ultralytics==8.1.0  # YOLOv8
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
sentence-transformers==2.2.2  # Face embeddings

# Face Recognition
face-recognition==1.3.0
deepface==0.0.79

# Data & Utils
numpy==1.26.3
pandas==2.1.4
python-dateutil==2.8.2
pyyaml==6.0.1
python-dotenv==1.0.0
jsonschema==4.20.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Async & Queue
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Data Generation
Faker==22.0.0
reportlab==4.0.9  # PDF generation

#QR
qrcode==7.4.2